add some functions, update kvm17275 17200,add hyperv17275, add rhevm17200
