+++ testcases/virt_who_polarion/hyperv/tc_ID17275_HYPERV_validate_compliance_unregister_host_check_mapping_when_re_register_host.py
@@ -1,23 +1,23 @@
 from utils import *
-from testcases.virt_who_polarion.kvmbase import KVMBase
+from testcases.virt_who_polarion.hypervbase import HYPERVBase
 from utils.exception.failexception import FailException
 
-class tc_ID17275_validate_compliance_unregister_host_check_mapping_when_re_register_host(KVMBase):
+class tc_ID17275_HYPERV_validate_compliance_unregister_host_check_mapping_when_re_register_host(HYPERVBase):
     def test_run(self):
         case_name = self.__class__.__name__
         logger.info("========== Begin of Running Test Case %s ==========" % case_name)
         try:
             SERVER_IP, SERVER_HOSTNAME, SERVER_USER, SERVER_PASS = self.get_server_info()
 
-            guest_name = self.get_vw_cons("KVM_GUEST_NAME")
-            guestuuid = self.vw_get_uuid(guest_name)
-
+            guest_name = self.get_vw_cons("HYPERV_GUEST_NAME")
             test_sku = self.get_vw_cons("productid_unlimited_guest")
             bonus_quantity = self.get_vw_cons("guestlimit_unlimited_guest")
             sku_name = self.get_vw_cons("productname_unlimited_guest")
 
-            self.vw_start_guests(guest_name)
-            guestip = self.kvm_get_guest_ip(guest_name)
+            self.hyperv_start_guest(guest_name)
+            guestip = self.hyperv_get_guest_ip(guest_name)
+            guestuuid = self.hyperv_get_guest_guid(guest_name)
+            hostuuid = self.hyperv_get_host_uuid()
 
             # register guest to SAM
             if not self.sub_isregistered(guestip):
@@ -25,21 +25,30 @@ class tc_ID17275_validate_compliance_unregister_host_check_mapping_when_re_regis
                 self.sub_register(SERVER_USER, SERVER_PASS, guestip)
             #(1) Validate guest consumed bonus pool will revoke after unregister host
             # subscribe the host to the physical pool which can generate bonus pool
-            self.sub_subscribe_sku(test_sku)
+            # host subscribe datacenter pool
+            self.server_subscribe_system(hostuuid, self.get_poolid_by_SKU(test_sku), SERVER_IP)
             # subscribe the registered guest to the corresponding bonus pool
             self.sub_subscribe_to_bonus_pool(test_sku, guestip)
             # list consumed subscriptions on guest
             self.sub_listconsumed(sku_name, guestip)
-            # unregister hosts
-            self.sub_unregister()
+            #(2) Check guest uuid after unregister host and hypervisor
+            tmp_file = "/tmp/tail.rhsm.log"
+            cmd = "subscription-manager unregister"
+            self.generate_tmp_log(cmd, tmp_file)
+            cmd = "cat %s" % tmp_file
+            ret, output = self.runcmd(cmd, "get temporary log generated")
+            if ret == 0 and "not registered" in output:
+                logger.info("Success to check virt-who log after unregister host")
+            else:
+                raise FailException("failed to check virt-who log after unregister host")
+            self.server_remove_system(hostuuid, SERVER_IP)
 #             time.sleep(60)
             self.sub_refresh(guestip)
             # list consumed subscriptions on guest
             self.sub_listconsumed(sku_name, guestip, productexists=False)
-
-            #(2). Check guest uuid after re-register
+            #(3). Check guest uuid after re-register
             self.sub_register(SERVER_USER, SERVER_PASS)
-            self.vw_check_uuid(guestuuid, uuidexists=True)
+            self.hypervisor_check_uuid(hostuuid, guestuuid, uuidexists=True)
 
             self.assert_(True, case_name)
         except Exception, e:
@@ -47,7 +56,7 @@ class tc_ID17275_validate_compliance_unregister_host_check_mapping_when_re_regis
             self.assert_(False, case_name)
         finally:
             self.sub_unregister(guestip)
-            self.vw_stop_guests(guest_name)
+            self.hyperv_stop_guest(guest_name)
             # register host
             self.sub_register(SERVER_USER, SERVER_PASS)
             logger.info("========== End of Running Test Case: %s ==========" % case_name)
